+++ middleware/redirect.go
@@ -9,48 +9,71 @@ import (
 )
 
 // httpsRedirector is a middleware that redirects http requests to https.
+// It also protects against [DNS rebinding] attacks.
 //
 // domain is the domain name of your website.
 // httpsPort is the tls port where http requests will be redirected to.
+//
+// [DNS rebinding]: https://en.wikipedia.org/wiki/DNS_rebinding
 func httpsRedirector(wrappedHandler http.Handler, httpsPort uint16, domain string) http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
-		isTls := strings.EqualFold(r.URL.Scheme, "https") || r.TLS != nil
-		if !isTls {
-			url := r.URL
-			url.Scheme = "https"
-			url.Host = joinHostPort(domain, fmt.Sprint(httpsPort))
-			path := url.String()
+		{ // 1. http -> https redirect.
+			isTls := strings.EqualFold(r.URL.Scheme, "https") || r.TLS != nil
+			if !isTls {
+				url := r.URL
+				url.Scheme = "https"
+				url.Host = joinHostPort(domain, fmt.Sprint(httpsPort))
+				path := url.String()
 
-			http.Redirect(w, r, path, http.StatusPermanentRedirect)
-			return
+				http.Redirect(w, r, path, http.StatusPermanentRedirect)
+				return
+			}
 		}
 
-		// A Host header field must be sent in all HTTP/1.1 request messages.
-		// Thus we expect `r.Host[0]` to always have a value.
-		// https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Host
-		isHostBareIP := unicode.IsDigit(rune(r.Host[0]))
-		if isHostBareIP {
-			/*
-				the request has tried to access us via an IP address, redirect them to our domain.
-
-				curl -vkIL 172.217.170.174 #google
-				HEAD / HTTP/1.1
-				Host: 172.217.170.174
-
-				HTTP/1.1 301 Moved Permanently
-				Location: http://www.google.com/
-			*/
-			url := r.URL
-			url.Scheme = "https"
-			_, port, err := net.SplitHostPort(r.Host)
-			if err != nil {
+		host, port := getHostPort(r.Host)
+		{ // 2. bareIP -> https redirect.
+			// A Host header field must be sent in all HTTP/1.1 request messages.
+			// Thus we expect `r.Host[0]` to always have a value.
+			// https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Host
+			if port == "" {
 				port = fmt.Sprint(httpsPort)
 			}
-			url.Host = joinHostPort(domain, port)
-			path := url.String()
+			isHostBareIP := unicode.IsDigit(rune(host[0]))
+			if isHostBareIP {
+				/*
+					the request has tried to access us via an IP address, redirect them to our domain.
 
-			http.Redirect(w, r, path, http.StatusPermanentRedirect)
-			return
+					curl -vkIL 172.217.170.174 #google
+					HEAD / HTTP/1.1
+					Host: 172.217.170.174
+
+					HTTP/1.1 301 Moved Permanently
+					Location: http://www.google.com/
+				*/
+				url := r.URL
+				url.Scheme = "https"
+				url.Host = joinHostPort(domain, port)
+				path := url.String()
+
+				http.Redirect(w, r, path, http.StatusPermanentRedirect)
+				return
+			}
+		}
+
+		{ // 3. DNS rebinding attack protection.
+			// todo: before calling `isDomainOrSubdomain` we need to make sure that both args are already be in canonical form.
+			// see; https://github.com/golang/go/blob/master/src/net/http/client.go#L1001-L1003
+			// We know that domain is kinda already canonical since [New] validates that. But host is not.
+			if !isDomainOrSubdomain(host, domain) {
+				err := fmt.Errorf("ong/middleware: the HOST http header has an unexpected value")
+				w.Header().Set(ongMiddlewareErrorHeader, err.Error())
+				http.Error(
+					w,
+					err.Error(),
+					http.StatusBadRequest,
+				)
+				return
+			}
 		}
 
 		wrappedHandler.ServeHTTP(w, r)
@@ -88,4 +111,39 @@ func indexByteString(s string, c byte) int {
 		}
 	}
 	return -1
+}
+
+// getHostPort returns host and port.
+// It is based on `http.stripHostPort` from https://github.com/golang/go/blob/go1.20.5/src/net/http/server.go#L2348-L2349
+func getHostPort(h string) (host, port string) {
+	// If no port on host, return unchanged
+	if !strings.Contains(h, ":") {
+		return h, ""
+	}
+
+	hst, prt, err := net.SplitHostPort(h)
+	if err != nil {
+		return h, prt // on error, return unchanged
+	}
+
+	return hst, prt
+}
+
+// isDomainOrSubdomain reports whether sub is a subdomain (or exact
+// match) of the parent domain.
+// Both domains must already be in canonical form.
+// // It is based on `http.isDomainOrSubdomain` from https://github.com/golang/go/blob/master/src/net/http/client.go#L1009-L1013
+func isDomainOrSubdomain(sub, parent string) bool {
+	if sub == parent {
+		return true
+	}
+
+	// If sub is "foo.example.com" and parent is "example.com",
+	// that means sub must end in "."+parent.
+	// Do it without allocating.
+	if !strings.HasSuffix(sub, parent) {
+		return false
+	}
+
+	return sub[len(sub)-len(parent)-1] == '.'
 }